refactor(m2): REF-206/207 apps 切换至 @pixuli/core + @pixuli/ui
pixuli 与 mobile 移除 @packages/common 路径；类型/工具/日志走 core，UI 走 ui（native 子路径）；存储服务暂经 pixuli-common/services。core 新增 image-browser 共享类型。
