+++ apps/mobile/package.json
@@ -39,6 +39,8 @@
     "expo-system-ui": "~6.0.9",
     "i18next": "^25.6.0",
     "octokit": "3.1.2",
+    "@pixuli/core": "workspace:*",
+    "@pixuli/ui": "workspace:*",
     "pixuli-common": "workspace:*",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
