+++ apps/pixuli/package.json
@@ -26,6 +26,8 @@
     "i18next-browser-languagedetector": "^8.2.0",
     "lucide-react": "^0.556.0",
     "octokit": "3.1.2",
+    "@pixuli/core": "workspace:*",
+    "@pixuli/ui": "workspace:*",
     "pixuli-common": "workspace:*",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
