+++ packages/common/package.json
@@ -2,9 +2,15 @@
   "name": "pixuli-common",
   "version": "1.0.0",
   "description": "Pixuli common",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./native": "./src/index.native.ts",
+    "./services": "./src/services/index.ts",
+    "./services/native": "./src/services/index.native.ts"
+  },
   "files": ["dist", "README.md", "src", "docs"],
   "publishConfig": {
     "access": "public"
