+++ packages/core/src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './image';
+export * from './image-browser';
 export * from './github';
 export * from './gitee';
 export * from './log';
