+++ packages/ui/src/image/image-browser/common/types.ts
@@ -1,14 +1,9 @@
 /**
- * ImageBrowser 组件共享类型定义
+ * ImageBrowser 组件共享类型（定义在 @pixuli/core，此处 re-export）
  */
-
-export interface FilterOptions {
-  searchTerm: string;
-  selectedTypes: string[];
-  selectedTags: string[];
-}
-
-export type SortField = 'createdAt' | 'name' | 'size';
-export type SortOrder = 'asc' | 'desc';
-
-export type ViewMode = 'grid' | 'list';
+export type {
+  FilterOptions,
+  SortField,
+  SortOrder,
+  ViewMode,
+} from '@pixuli/core/types';
